requirements.txt
ipython==7.6.1
numpy
Pillow
scipy==1.1
numpy
tensorflow-gpu >= 1.0